# Set DB pool size for all Crystal impelementations