--- a/frameworks/Crystal/crystal/crystal.dockerfile
+++ b/frameworks/Crystal/crystal/crystal.dockerfile
@@ -8,6 +8,7 @@ COPY shard.lock shard.lock
 COPY shard.yml shard.yml
 
 ENV GC_MARKERS 1
+ENV DATABASE_URL postgres://benchmarkdbuser:benchmarkdbpass@tfb-database:5432/hello_world?initial_pool_size=8&max_pool_size=8&max_idle_pool_size=8
 
 RUN shards install
 RUN crystal build --release --no-debug server.cr -o server.out