--- a/frameworks/Crystal/kemal/kemal.dockerfile
+++ b/frameworks/Crystal/kemal/kemal.dockerfile
@@ -9,6 +9,7 @@ COPY shard.yml shard.yml
 
 ENV GC_MARKERS 1
 ENV KEMAL_ENV production
+ENV DATABASE_URL postgres://benchmarkdbuser:benchmarkdbpass@tfb-database:5432/hello_world?initial_pool_size=8&max_pool_size=8&max_idle_pool_size=8
 
 RUN shards install
 RUN crystal build --release --no-debug server-postgres.cr