--- a/frameworks/Crystal/raze/raze.dockerfile
+++ b/frameworks/Crystal/raze/raze.dockerfile
@@ -7,6 +7,8 @@ COPY raze.cr raze.cr
 COPY shard.lock shard.lock
 COPY shard.yml shard.yml
 
+ENV DATABASE_URL postgres://benchmarkdbuser:benchmarkdbpass@tfb-database:5432/hello_world?initial_pool_size=8&max_pool_size=8&max_idle_pool_size=8
+
 RUN shards install
 RUN crystal build --release --no-debug raze.cr