Lower DB pool size from #4160 results in better RPS on Citrine.

Before https://www.techempower.com/benchmarks/#section=test&runid=56076e97-0658-46a1-81bb-6f8890f2e85e
After https://www.techempower.com/benchmarks/#section=test&runid=5191fd86-6a2a-405b-b326-f160fc6781b7
```
                  Before    After    Better
JSON              345k   -> 1016k  ~ 3x
Single query      178k   -> 261k   ~ 1.5x
Multiple queries  9.2k   -> 11.9k  ~ 1.3x
Fortunes          87.5k  -> 258.4k ~ 3x
Data updates      1.2k   -> 3.1k   ~ 2.7x
Plain Text        740k   -> 2025k  ~ 2.7x
```
Note:

Keep an eye on Latency, since it's dropped a lot in some cases.

Might be a good idea to try average latency from multiple runs, as latency numbers seem to be in flux sometimes for no obvious reason.
```
                  Before    After    Worse
JSON              0.2ms  -> 0.7ms  ~ 3x
Single query      0.6ms  -> 0.4 ms
Multiple queries  55ms   -> 309ms  > 5x
Fortunes          1.3ms  -> 173ms  > 100x (!!!)
Data updates      426sm  -> 535ms  ~ 1.2x
Plain Text        4.3ms  -> 13.9ms ~ 3x
```
<!--
Thank you for submitting to the TechEmpower Framework Benchmarks!

If you are submitting a new framework, please make sure that you add the appropriate line in the `.travis.yml` file for proper integration testing. Also please make sure that an appropriate `README.md` is added in your framework directory with information about the framework and a link to its homepage and documentation.

If you are editing an existing test, please update the `README.md` for that test where appropriate.
-->